gh-130931: Add pycore_interpframe.h internal header
Move _PyInterpreterFrame and associated functions
to a new pycore_interpframe.h header.
